app/requirements.txt
tensorflow==2.3.0
Pillow==7.2.0
gunicorn>=19.7.1
flask
redis